[상진규] fix : ProductCard 수정
