+++ src/components/common/ProductGrid.tsx
@@ -4,8 +4,8 @@ import ProductCard from '@/components/common/ProductCard';
 const ProductGrid = ({ title, products }: ProductGridProps) => {
   return (
     <>
-      <h1 className='text-[#F1F1F5] lg:text-2xl text-xl font-semibold'>{title}</h1>
-      <div className='grid grid-cols-2 lg:grid-cols-3 gap-[15px] lg:gap-5 justify-center'>
+      <h1 className='text-[#F1F1F5] text-xl font-semibold'>{title}</h1>
+      <div className='grid grid-cols-2 gap-[15px] lg:gap-[20px] lg:grid-cols-3'>
         {products.map((product) => (
           <ProductCard key={product.id} product={product} />
         ))}
